- Other: Remove all AWS specific logic out of blueair-api.ts, and into blueair-aws-api.ts; refactor all necesssary files to repoint references.
