+++ src/platform.ts
@@ -1,6 +1,7 @@
 import { API, DynamicPlatformPlugin, Logger, PlatformAccessory, PlatformConfig, Service, Characteristic } from 'homebridge';
 
 import { BlueAirApi } from './blueair-api';
+import { BlueAirAwsApi } from './blueair-aws-api';
 import { PLATFORM_NAME, PLUGIN_NAME } from './settings';
 
 import { BlueAirPlatformAccessory } from './platformAccessory';
@@ -21,6 +22,7 @@ export class BlueAirHomebridgePlatform implements DynamicPlatformPlugin {
   public readonly accessories: PlatformAccessory[] = [];
 
   readonly blueair!: BlueAirApi;
+  readonly blueairAws!: BlueAirAwsApi;
 
   constructor(
     public readonly log: Logger,
@@ -35,7 +37,7 @@ export class BlueAirHomebridgePlatform implements DynamicPlatformPlugin {
       return;
     }
 
-    this.blueair = new BlueAirApi(this.log, config.username, config.password, config.region);
+    this.blueairAws = new BlueAirAwsApi(this.log, config.username, config.password, config.region);
 
     this.log.debug('Finished initializing platform:', this.config.name);
 
@@ -157,21 +159,21 @@ export class BlueAirHomebridgePlatform implements DynamicPlatformPlugin {
   async discoverAwsDevices() {
 
     // login to BlueAir
-    const login_flag: boolean = await this.blueair.awsLogin();
+    const login_flag: boolean = await this.blueairAws.awsLogin();
     if(!login_flag){
       this.log.error('Failed to login to AWS. Check password and restart Homebridge to try again.');
       return false;
     }
 
     // retrieve devices
-    const devices_flag = await this.blueair.getAwsDevices();
+    const devices_flag = await this.blueairAws.getAwsDevices();
     if(!devices_flag){
       this.log.error('Failed to get list of AWS devices. Check BlueAir App.');
       return false;
     }
 
     // loop over the discovered devices and register each one if it has not already been registered
-    for (const device of this.blueair.awsDevices) {
+    for (const device of this.blueairAws.awsDevices) {
       // generate a unique id for the accessory this should be generated from
       // something globally unique, but constant, for example, the device serial
       // number or MAC address
@@ -206,7 +208,7 @@ export class BlueAirHomebridgePlatform implements DynamicPlatformPlugin {
           continue;
         }
 
-        const deviceInfo = await this.blueair.getAwsDeviceInfo(device.name, device.uuid);
+        const deviceInfo = await this.blueairAws.getAwsDeviceInfo(device.name, device.uuid);
         //this.log.info('Device Info:', deviceInfo);
 
         this.log.info('Adding new accessory:', device.name);
@@ -319,7 +321,7 @@ export class BlueAirHomebridgePlatform implements DynamicPlatformPlugin {
   // AWS Accessory currently handles DustMagnet and Health Protect
   private async findAwsModelAndInitialize(device, accessory){
     // retreive model info
-    const info = await this.blueair.getAwsDeviceInfo(device.name, device.uuid);
+    const info = await this.blueairAws.getAwsDeviceInfo(device.name, device.uuid);
     this.log.debug('Device Info from findAwsModelAndInitialize: ', info);
     //this.log.info('%s of type "%s" initialized.', device.configuration.di.name, info.compatibility);
